fix error search function on android 7
